+++ sysl2/sysl/parse/tests/absolute_import.sysl
@@ -1,0 +1,4 @@
+import /alias
+
+Model:
+    !alias FooBar : string
